requirements.txt
adal==1.2.7
annotated-types==0.7.0
anyio==3.7.1
azure-batch==13.0.0
azure-common==1.1.28
azure-core==1.32.0
azure-servicebus==7.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.104.1
h11==0.14.0
idna==3.10
isodate==0.7.2
msrest==0.7.1
msrestazure==0.6.4.post1
oauthlib==3.2.2
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
websockets==12.0
sqlalchemy==2.0.27
asyncpg==0.29.0
